requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
python-dotenv==1.0.1
pydantic==2.8.2
email-validator>=2.0.0
pytest==8.3.2